Make it clear that config.ts is for Cnd and move it under environment/
