+++ tests/src/environment/cnd_instance.ts
@@ -1,7 +1,7 @@
 import { JsonMap, stringify } from "@iarna/toml";
 import { ChildProcess, spawn } from "child_process";
 import tempWrite from "temp-write";
-import { CndConfigFile } from "../config";
+import { CndConfigFile } from "./cnd_config_file";
 import { sleep } from "../utils";
 import waitForLogMessage from "./wait_for_log_message";
 import { Logger } from "log4js";
